minor updates
gameinclude.h has 2 new includes vector and memory.  The vector works
like an array list so we can make multiple units of x class,
the NPCUnits class is where we define our units like cells to be
attacked by the bacteria, or white blood cells
the scatterShotSkill is the influenza bacteria ability, apprently it
would be easier to create each skill as a class instead of putting it
into the player class given the amount of data required for it.